Refactored folder split index out into seperate files Board, player, & tilesetGame
